+++ mne/montages/__init__.py
@@ -1,3 +1,0 @@
-"""Module to handle EEG montages"""
-
-from .montage import read_montage, apply_montage