+++ mne/viz/topomap.py
@@ -29,7 +29,7 @@ def _prepare_topo_plot(obj, ch_type, layout):
     """"Aux Function"""
     info = copy.deepcopy(obj.info)
     if layout is None and ch_type is not 'eeg':
-        from ..layouts.layout import find_layout
+        from ..channels import find_layout
         layout = find_layout(info)
     elif layout == 'auto':
         layout = None
@@ -41,7 +41,7 @@ def _prepare_topo_plot(obj, ch_type, layout):
     # special case for merging grad channels
     if (ch_type == 'grad' and FIFF.FIFFV_COIL_VV_PLANAR_T1 in
             np.unique([ch['coil_type'] for ch in info['chs']])):
-        from ..layouts.layout import _pair_grad_sensors
+        from ..channels.layout  import _pair_grad_sensors
         picks, pos = _pair_grad_sensors(info, layout)
         merge_grads = True
     else:
@@ -58,7 +58,7 @@ def _prepare_topo_plot(obj, ch_type, layout):
 
         if layout is None:
             chs = [info['chs'][i] for i in picks]
-            from ..layouts.layout import _find_topomap_coords
+            from ..channels.layout  import _find_topomap_coords
             pos = _find_topomap_coords(chs, layout)
         else:
             names = [n.upper() for n in layout.names]
@@ -88,7 +88,7 @@ def _plot_update_evoked_topomap(params, bools):
     data = new_evoked.data[np.ix_(params['picks'],
                                   params['time_idx'])] * params['scale']
     if params['merge_grads']:
-        from ..layouts.layout import _merge_grad_data
+        from ..channels.layout  import _merge_grad_data
         data = _merge_grad_data(data)
     image_mask = params['image_mask']
 
@@ -154,7 +154,7 @@ def plot_projs_topomap(projs, layout=None, cmap='RdBu_r', sensors=True,
     import matplotlib.pyplot as plt
 
     if layout is None:
-        from ..layouts import read_layout
+        from ..channels import read_layout
         layout = read_layout('Vectorview-all')
 
     if not isinstance(layout, list):
@@ -175,7 +175,7 @@ def plot_projs_topomap(projs, layout=None, cmap='RdBu_r', sensors=True,
         for l in layout:
             is_vv = l.kind.startswith('Vectorview')
             if is_vv:
-                from ..layouts.layout import _pair_grad_sensors_from_ch_names
+                from ..channels.layout  import _pair_grad_sensors_from_ch_names
                 grad_pairs = _pair_grad_sensors_from_ch_names(ch_names)
                 if grad_pairs:
                     ch_names = [ch_names[i] for i in grad_pairs]
@@ -186,7 +186,7 @@ def plot_projs_topomap(projs, layout=None, cmap='RdBu_r', sensors=True,
 
             pos = l.pos[idx]
             if is_vv and grad_pairs:
-                from ..layouts.layout import _merge_grad_data
+                from ..channels.layout  import _merge_grad_data
                 shape = (len(idx) / 2, 2, -1)
                 pos = pos.reshape(shape).mean(axis=1)
                 data = _merge_grad_data(data[grad_pairs]).ravel()
@@ -644,7 +644,7 @@ def plot_ica_components(ica, picks=None, ch_type='mag', res=64,
     fig.suptitle(title)
 
     if merge_grads:
-        from ..layouts.layout import _merge_grad_data
+        from ..channels.layout  import _merge_grad_data
     for ii, data_, ax in zip(picks, data, axes):
         ax.set_title('IC #%03d' % ii, fontsize=12)
         data_ = _merge_grad_data(data_) if merge_grads else data_
@@ -793,7 +793,7 @@ def plot_tfr_topomap(tfr, tmin=None, tmax=None, fmin=None, fmax=None,
     data = np.mean(np.mean(data, axis=2), axis=1)[:, np.newaxis]
 
     if merge_grads:
-        from ..layouts.layout import _merge_grad_data
+        from ..channels.layout  import _merge_grad_data
         data = _merge_grad_data(data)
 
     vmin, vmax = _setup_vmin_vmax(data, vmin, vmax)
@@ -998,7 +998,7 @@ def plot_evoked_topomap(evoked, times=None, ch_type='mag', layout=None,
 
     data *= scale
     if merge_grads:
-        from ..layouts.layout import _merge_grad_data
+        from ..channels.layout  import _merge_grad_data
         data = _merge_grad_data(data)
 
     vmin, vmax = _setup_vmin_vmax(data, vmin, vmax)